% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitgantt.R
\name{gg_extract_dates}
\alias{gg_extract_dates}
\alias{gg_start_date}
\alias{gg_due_date}
\title{Extract dates from gh issues}
\usage{
gg_extract_dates(issue_body, pattern = "GanttStart: |GanttDue: ")

gg_start_date(issue_body)

gg_due_date(issue_body)
}
\arguments{
\item{issue_body}{The body of an issue that contains `GanttStart` and `GanttDue` lines}

\item{pattern}{The text string used to identify start and due dates}
}
\description{
Extract dates from gh issues
}
\examples{
issue_body = "GanttStart: 2019-09-01\\r\\nGanttDue: 2019-10-01"
gg_extract_dates(issue_body)
gg_start_date(issue_body)
gg_due_date(issue_body)
}
