% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitgantt.R
\name{gg_issue_list}
\alias{gg_issue_list}
\title{Create a list of issues in a GitHub repo}
\usage{
gg_issue_list(owner, repo, state = "all", limit = Inf)
}
\arguments{
\item{owner}{The owner of the repo, e.g. `ropensci`}

\item{repo}{The name of the repo, e.g. `gh`}

\item{state}{One of `"all"` (the default), `"open"`, or `"closed"`}

\item{limit}{Maximum number of issues to return, `Inf` by default}
}
\description{
Create a list of issues in a GitHub repo
}
\examples{
owner = "ATFutures"
repo = "who3"
gg_issue_list(owner, repo, limit = 2)
}
