% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitgantt.R
\name{gg_timevis}
\alias{gg_timevis}
\title{Create interactive visualisation of GitHub issues}
\usage{
gg_timevis(issue_df = NULL, owner, repo)
}
\arguments{
\item{issue_df}{A data frame created by `gg_issue_df()`}

\item{owner}{The owner of the repo, e.g. `ropensci`}

\item{repo}{The name of the repo, e.g. `gh`}
}
\description{
Create interactive visualisation of GitHub issues
}
\examples{
issue_df = gg_issue_df(gg_issue_list("ATFutures", "who3"))
gg_timevis(issue_df)
gg_timevis(owner = "ATFutures", repo = "who3")
}
